requirements.txt
fastapi
uvicorn
asyncpg
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
websockets